chore: update implementation
